feat: add tabs for filter sets
